chore: Update .gitignore and webpack.config.js
This commit updates the .gitignore file to exclude *.zip files from version control. It also modifies the webpack.config.js file to change the output path for the generated bundles. The bundles will now be placed in the "./build/" directory instead of "./build/www/". These changes aim to improve the project's build process and ensure that unnecessary files are not included in the repository.